USB: UAS: fix disconnect by unplugging a hub
mainline inclusion
from mainline-v5.9-rc5
commit eb2a86ae8c544be0ab04aa8169390c0669bc7148
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I9W2CC
CVE: NA

-------------------

The SCSI layer can go into an ugly loop if you ignore that a device is
gone. You need to report an error in the command rather than in the
return value of the queue method.

We need to specifically check for ENODEV. The issue goes back to the
introduction of the driver.

Fixes: 115bb1ffa54c3 ("USB: Add UAS driver")
Signed-off-by: Oliver Neukum <oneukum@suse.com>
Link: https://lore.kernel.org/r/20200916094026.30085-2-oneukum@suse.com
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
Signed-off-by: leoliu-oc <leoliu-oc@zhaoxin.com>