KVM: x86/mmu: allow APICv memslot to be enabled but invisible
mainline inclusion
from mainline-v5.15-rc1
commit 9cc13d60ba6b9976e01ec6f66fa1ec4a06992929
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IBTVRA

----------------------------------------

commit 9cc13d60ba6b9976e01ec6f66fa1ec4a06992929 upstream

on AMD, APIC virtualization needs to dynamicaly inhibit the AVIC in a
response to some events, and this is problematic and not efficient to do by
enabling/disabling the memslot that covers APIC's mmio range.

Plus due to SRCU locking, it makes it more complex to
request AVIC inhibition.

Instead, the APIC memslot will be always enabled, but be invisible
to the guest, such as the MMU code will not install a SPTE for it,
when it is inhibited and instead jump straight to emulating the access.

When inhibiting the AVIC, this SPTE will be zapped.

This code is based on a suggestion from Sean Christopherson:
https://lkml.org/lkml/2021/7/19/2970

[Backport Changes]

1. Fixing a22842acf6bb and 664a04f37a2d,
In arch/x86/kvm/mmu/mmu.c, if checks related to "private memslots to L2"
in the static function kvm_faultin_pfn() were already updated with the
latest upstream changes by the Euler community during earlier backports.
However, during this backport, changes to the if block handling APIC
access page private memslots with the "APIC_ACCESS_PAGE_PRIVATE_MEMSLOT"
were skipped. This commit corrects that by adding the required if blocks
with up-to-date changes, ensuring full alignment with the earlier
backported upstream commit 5276c616abf1 and the latest upstream code.

Suggested-by: Sean Christopherson <seanjc@google.com>
Signed-off-by: Maxim Levitsky <mlevitsk@redhat.com>
Message-Id: <20210810205251.424103-8-mlevitsk@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: PvsNarasimha <PVS.NarasimhaRao@amd.com>
Signed-off-by: mohanasv <mohanasv@amd.com>