KVM: SVM: Don't bother checking for "running" AVIC when kicking for IPIs
mainline inclusion
from mainline-v5.17-rc1
commit 202470d536b2cad22fa859f3e01202571c49ded9
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IBTVRA

----------------------------------------

commit 202470d536b2cad22fa859f3e01202571c49ded9 upstream

Drop the avic_vcpu_is_running() check when waking vCPUs in response to a
VM-Exit due to incomplete IPI delivery.  The check isn't wrong per se, but
it's not 100% accurate in the sense that it doesn't guarantee that the vCPU
was one of the vCPUs that didn't receive the IPI.

The check isn't required for correctness as blocking == !running in this
context.

From a performance perspective, waking a live task is not expensive as the
only moderately costly operation is a locked operation to temporarily
disable preemption.  And if that is indeed a performance issue,
kvm_vcpu_is_blocking() would be a better check than poking into the AVIC.

Signed-off-by: Sean Christopherson <seanjc@google.com>
Reviewed-by: Maxim Levitsky <mlevitsk@redhat.com>
Message-Id: <20211208015236.1616697-12-seanjc@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: chaithanyaLagisetty <LagisettyVenkata.NagaChaitanya@amd.com>
Signed-off-by: mohanasv <mohanasv@amd.com>