KVM: SVM: Use kvm_vcpu_is_blocking() in AVIC load to handle preemption
mainline inclusion
from mainline-v5.17-rc1
commit af52f5aa5c1b46809834b728a13a1af5aab451e9
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IBTVRA

----------------------------------------

commit af52f5aa5c1b46809834b728a13a1af5aab451e9 upstream

Use kvm_vcpu_is_blocking() to determine whether or not the vCPU should be
marked running during avic_vcpu_load().  Drop avic_is_running, which
really should have been named "vcpu_is_not_blocking", as it tracked if
the vCPU was blocking, not if it was actually running, e.g. it was set
during svm_create_vcpu() when the vCPU was obviously not running.

This is technically a teeny tiny functional change, as the vCPU will be
marked IsRunning=1 on being reloaded if the vCPU is preempted between
svm_vcpu_blocking() and prepare_to_rcuwait().  But that's a benign change
as the vCPU will be marked IsRunning=0 when KVM voluntarily schedules out
the vCPU.

Signed-off-by: Sean Christopherson <seanjc@google.com>
Message-Id: <20211208015236.1616697-14-seanjc@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: chaithanyaLagisetty <LagisettyVenkata.NagaChaitanya@amd.com>
Signed-off-by: mohanasv <mohanasv@amd.com>