arm64: RME: Allow VMM to set RIPAS
community inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IBY08N

Reference: https://lore.kernel.org/kvm/yq5amsgsrzva.fsf@kernel.org/T/#meb4cb12031992f1275fe7f5b40ac38c1e883ac87

--------------------------------

Each page within the protected region of the realm guest can be marked
as either RAM or EMPTY. Allow the VMM to control this before the guest
has started and provide the equivalent functions to change this (with
the guest's approval) at runtime.

When transitioning from RIPAS RAM (1) to RIPAS EMPTY (0) the memory is
unmapped from the guest and undelegated allowing the memory to be reused
by the host. When transitioning to RIPAS RAM the actual population of
the leaf RTTs is done later on stage 2 fault, however it may be
necessary to allocate additional RTTs to allow the RMM track the RIPAS
for the requested range.

When freeing a block mapping it is necessary to temporarily unfold the
RTT which requires delegating an extra page to the RMM, this page can
then be recovered once the contents of the block mapping have been
freed.

Signed-off-by: Steven Price <steven.price@arm.com>

Conflicts:
	arch/arm64/kvm/mmu.c
	arch/arm64/kvm/rme.c
[zhuangyiwei: keep using unmap_stage2_range instead of kvm_stage2_unmap_range which is modifed by nv features]
[zhuangyiwei: add may_block parameter to unmap_stage2_range instead of introducing bunch of nv patches]
[zhuangyiwei: let may_block=true for all unmap_stage2_range calls as original unmap_stage2_range calls' default value]
[zhuangyiwei: use get_page on undelegate failure instead of put_page on undelegate success because of missing guest_memfd]
Signed-off-by: Yiwei Zhuang <zhuangyiwei@huawei.com>