KVM: SVM: Document that vCPU ID == APIC ID in AVIC kick fastpatch
mainline inclusion
from mainline-v6.3-rc1
commit 8578e4512d873b0aa9805e9772162578d54bb6a1
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IBTVRA

----------------------------------------

commit 8578e4512d873b0aa9805e9772162578d54bb6a1 upstream

Document that AVIC is inhibited if any vCPU's APIC ID diverges from its
vCPU ID, i.e. that there's no need to check for a destination match in
the AVIC kick fast path.

Opportunistically tweak comments to remove "guest bug", as that suggests
KVM is punting on error handling, which is not the case.  Targeting a
non-existent vCPU or no vCPUs _may_ be a guest software bug, but whether
or not it's a guest bug is irrelevant.  Such behavior is architecturally
legal and thus needs to faithfully emulated by KVM (and it is).

Signed-off-by: Sean Christopherson <seanjc@google.com>
Reviewed-by: Maxim Levitsky <mlevitsk@redhat.com>
Message-Id: <20230106011306.85230-16-seanjc@google.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Arukonda Rahul <arukonda.rahul@amd.com>
Signed-off-by: mohanasv <mohanasv@amd.com>