KVM: SEV: store VMSA features in kvm_sev_info
mainline inclusion
from mainline-v6.10-rc1
commit 605bbdc12bc8839b97f8755d74fc3fd4bf206588
category: Feature
bugzilla: https://gitee.com/openeuler/kernel/issues/ICQGHL
CVE: NA

Reference: https://github.com/torvalds/linux/commit/605bbdc12bc8839b97f8755d74fc3fd4bf206588

--------------------------------

commit 605bbdc12bc8839b97f8755d74fc3fd4bf206588 upstream

Right now, the set of features that are stored in the VMSA upon
initialization is fixed and depends on the module parameters for
kvm-amd.ko.  However, the hypervisor cannot really change it at will
because the feature word has to match between the hypervisor and whatever
computes a measurement of the VMSA for attestation purposes.

Add a field to kvm_sev_info that holds the set of features to be stored
in the VMSA; and query it instead of referring to the module parameters.

Because KVM_SEV_INIT and KVM_SEV_ES_INIT accept no parameters, this
does not yet introduce any functional change, but it paves the way for
an API that allows customization of the features per-VM.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Message-Id: <20240209183743.22030-6-pbonzini@redhat.com>
Reviewed-by: Michael Roth <michael.roth@amd.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Message-ID: <20240404121327.3107131-7-pbonzini@redhat.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Jeevan deep J <j.jeevandeep@amd.com>
Signed-off-by: PrithivishS <sprithiv@amd.com>