KVM: guest_memfd: limit overzealous WARN
mainline inclusion
from mainline-v6.11-rc1
commit fa30b0dc91c815b9579d6f758437c35db059f5ae
category: Feature
bugzilla: https://gitee.com/openeuler/kernel/issues/ICQGHL
CVE: NA

Reference: https://github.com/torvalds/linux/commit/fa30b0dc91c815b9579d6f758437c35db059f5ae

--------------------------------

commit fa30b0dc91c815b9579d6f758437c35db059f5ae upstream

Because kvm_gmem_get_pfn() is called from the page fault path without
any of the slots_lock, filemap lock or mmu_lock taken, it is
possible for it to race with kvm_gmem_unbind().  This is not a
problem, as any PTE that is installed temporarily will be zapped
before the guest has the occasion to run.

However, it is not possible to have a complete unbind+bind
racing with the page fault, because deleting the memslot
will call synchronize_srcu_expedited() and wait for the
page fault to be resolved.  Thus, we can still warn if
the file is there and is not the one we expect.

Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: PrithivishS <sprithiv@amd.com>