x86/cpu/vfm: Add/initialize x86_vfm field to struct cpuinfo_x86
mainline inclusion
from mainline-v6.10-rc1
commit a9d0adce69075192961f3be466c4810a21b7bc9e
bugzilla: https://gitee.com/openeuler/kernel/issues/ICSPRX

----------------------------------------

commit a9d0adce69075192961f3be466c4810a21b7bc9e upstream

Refactor struct cpuinfo_x86 so that the vendor, family, and model
fields are overlaid in a union with a 32-bit field that combines
all three (together with a one byte reserved field in the upper
byte).

This will make it easy, cheap, and reliable to check all three
values at once.

See

  https://lore.kernel.org/r/Zgr6kT8oULbnmEXx@agluck-desk3

for why the ordering is (low-to-high bits):

  (vendor, family, model)

  [ bp: Move comments over the line, add the backstory about the
    particular order of the fields. ]

Signed-off-by: Tony Luck <tony.luck@intel.com>
Signed-off-by: Borislav Petkov (AMD) <bp@alien8.de>
Reviewed-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lore.kernel.org/r/20240416211941.9369-2-tony.luck@intel.com
Signed-off-by: Abhishek Rajput <abhishek.rajput@amd.com>
Signed-off-by: mohanasv2 <mohanasv@amd.com>