sw64: fix bugs in decreasing frequency
Sunway inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IBDJNZ

--------------------------------

The issue arose because the downshift_freq() was attempting to scale down
the frequency of a physicial core when only one of its logical core was
offline. The correct behavior should be to scale down the frequency only
when both logical cores of a pyhsicial core are offline.

Signed-off-by: Tang Jinyang <tangjinyang@wxiat.com>
Reviewed-by: He Sheng <hesheng@wxiat.com>
Signed-off-by: Gu Zitao <guzitao@wxiat.com>